pyproject.toml
[tool.poetry]
name = "airbyte-cortex-service"
version = "0.1.0"
description = "AI-ready data pre-processing for Airbyte."
authors = ["Kai <ngominhkhoa2006@gmail.com>"]
readme = "README.md"
packages = [
    {include = "cortex_service"}
]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
unstructured = {extras = ["local-inference"], version = "^0.18.14"}
sentence-transformers = ">=5.1.0,<6.0.0"
scikit-learn = ">=1.7.1,<2.0.0"
torch = ">=2.8.0,<3.0.0"
numpy = ">=2.3.2,<3.0.0"
fastapi = ">=0.116.1,<0.117.0"
uvicorn = ">=0.35.0,<0.36.0"
pydantic = {extras = ["email"], version = ">=2.11.7,<3.0.0"}
python-dotenv = ">=1.1.1,<2.0.0"
prometheus-fastapi-instrumentator = "^7.1.0"
python-json-logger = "^3.3.0"
requests = "^2.32.5"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
httpx = "^0.28.1"
ruff = "^0.12.11"
mypy = "^1.17.1"
black = "^25.1.0"

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
